make dem_id optional parameter in processing request payload
